Parameter Make.1-Machine

type 'a t

the machine computation

type 'a m

an external monad in which the machine computation is wrapped

module Observation : sig ... end

Observations interface.

module Syntax : sig ... end

Computation Syntax.

include Monads.Std.Monad.State.Multi.S with type 'a t := 'a t and type 'a m := 'a m and type env := Bap.Std.project and type id := Machine.id and module Syntax := Syntax and type 'a e = ?⁠boot:unit t -> ?⁠init:unit t -> ?⁠fini:unit t -> (exit_status * Bap.Std.project) m effect

include Monads.Std.Monad.Trans.S

type 'a t

type 'a m

type 'a e

val lift : 'a m -> 'a t

lifts inner monad into the resulting monad

val run : 'a t -> 'a e

runs the computation

type id

module Id : Core_kernel.Identifiable.S with type Id.t = id

val global : id

the identifier of the global (initial) state.

val fork : unit -> unit t

fork () forks the current state.

val switch : id -> unit t

switch id switches to the state with the given id if such state is alive, otherwise switches to the closest alive ancestor of the state with the given id

val parent : unit -> id t

parent () returns an identifier of the closest alive parent.

val ancestor : id list -> id t

ancestor ids returns an identifier of the closest common ancestor of states with the given identifiers.

val current : unit -> id t

current id returns an identifier of current state.

val kill : id -> unit t

kill id kills a state with the specified id. If id corresponds to the current state, then switches to the closest ancestor. If id = global then do nothing.

val forks : unit -> id Core_kernel.Sequence.t t

forks xs returns a sequence of all alive states

val status : id -> Monads.Std.Monad.State.Multi.status t

status id returns a status of a state with the given id

include Monads.Std.Monad.State.S with type 'a t := 'a t and type 'a e := 'a e and type 'a m := 'a m

include Monads.Std.Monad.Trans.S

type 'a t

type 'a m

type 'a e

val lift : 'a m -> 'a t

lifts inner monad into the resulting monad

val run : 'a t -> 'a e

runs the computation

include Monads.Std.Monad.Monad with type 'a t := 'a t

type 'a t

val void : 'a t -> unit t

void m computes m and discrards the result.

val sequence : unit t list -> unit t

sequence xs computes a sequence of computations xs in the left to right order.

val forever : 'a t -> 'b t

forever xs creates a computationt that never returns.

module Fn : sig ... end

Various function combinators lifted into the Kleisli category.

module Pair : sig ... end

The pair interface lifted into the monad.

module Triple : sig ... end

The triple interface lifted into a monad.

module Lift : sig ... end

Lifts functions into the monad.

module Exn : sig ... end

Interacting between monads and language exceptions

module Collection : sig ... end

Lifts collection interface into the monad.

module List : Collection.S with type 'a t := 'a list

The Monad.Collection.S interface for lists

module Seq : Collection.S with type 'a t := 'a Core_kernel.Sequence.t

The Monad.Collection.S interface for sequences

include Monads.Std.Monad.Syntax.S with type 'a t := 'a t

type 'a t

val (>>=) : 'a t -> ('a -> 'b t) -> 'b t

m >>= f is bind m f

val (>>|) : 'a t -> ('a -> 'b) -> 'b t

m >>= f is map m ~f

val (>=>) : ('a -> 'b t) -> ('b -> 'c t) -> 'a -> 'c t

f >=> g is fun x -> f x >>= g

val (!!) : 'a -> 'a t

!!x is return x

val (!$) : ('a -> 'b) -> 'a t -> 'b t

!$f is Lift.unary f

val (!$$) : ('a -> 'b -> 'c) -> 'a t -> 'b t -> 'c t

!$$f is Lift.binary f

val (!$$$) : ('a -> 'b -> 'c -> 'd) -> 'a t -> 'b t -> 'c t -> 'd t

!$$$f is Lift.ternary f

val (!$$$$) : ('a -> 'b -> 'c -> 'd -> 'e) -> 'a t -> 'b t -> 'c t -> 'd t -> 'e t

!$$$$f is Lift.quaternary f

val (!$$$$$) : ('a -> 'b -> 'c -> 'd -> 'e -> 'f) -> 'a t -> 'b t -> 'c t -> 'd t -> 'e t -> 'f t

!$$$$$f is Lift.quinary f

include Monads.Std.Monad.Syntax.Let.S with type 'a t := 'a t

type 'a t

val let* : 'a t -> ('a -> 'b t) -> 'b t

let* r = f x in b is f x >>= fun r -> b

val and* : 'a t -> 'b t -> ('a * 'b) t

monoidal product

val let+ : 'a t -> ('a -> 'b) -> 'b t

let+ r = f x in b is f x >>| fun r -> b

val and+ : 'a t -> 'b t -> ('a * 'b) t

monoidal product

include Core_kernel.Monad.S with type 'a Machine.t := 'a t

val (>>=) : 'a t -> ('a -> 'b t) -> 'b t

val (>>|) : 'a t -> ('a -> 'b) -> 'b t

module Monad_infix : sig ... end

val bind : 'a t -> f:('a -> 'b t) -> 'b t

val return : 'a -> 'a t

val map : 'a t -> f:('a -> 'b) -> 'b t

val join : 'a t t -> 'a t

val ignore_m : 'a t -> unit t

val all : 'a t list -> 'a list t

val all_unit : unit t list -> unit t

module Let_syntax : sig ... end

module Let : Monads.Std.Monad.Syntax.Let.S with type 'a t := 'a t

Monadic operators, see Monad.Syntax.S for more.

module Syntax : Monads.Std.Monad.Syntax.S with type 'a t := 'a t

Monadic operators, see Monad.Syntax.S for more.

type env

the environment type.

val put : env -> unit t

put s changes the current state to s

val get : unit -> env t

get s gets the current state

val gets : (env -> 'r) -> 'r t

gets p projects the current state with the function p

val update : (env -> env) -> unit t

update f updates the current state with the function f

module Local : Machine.State with type 'a m := 'a t and type 'a t := 'a Machine.state

Local state of the machine.

module Global : Machine.State with type 'a m := 'a t and type 'a t := 'a Machine.state

Global state shared across all machine clones.

module Other : sig ... end

Local state of other machines.

val raise : exn -> 'a t

raise exn raises the machine exception exn, intiating an abonormal control flow

val catch : 'a t -> (exn -> 'a t) -> 'a t

catch x f creates a computation that is equal to x if it terminates normally, and to f e if x terminates abnormally with the exception e.

val project : Bap.Std.project t

project is a computation that results with the project data structure. Note, that Machine is a State monad with the env type equal to project, thus project is a shortcut to get ().

This function is always evaluated in the global context, i.e., there is only one project that is shared by all machine forks.

You can use put project to update the project data structure.

val program : Bap.Std.program Bap.Std.term t

program program representation.

The same as gets Project.program.

val arch : Bap.Std.arch t

arch code architecture.

The same as gets Project.arch.

val args : string array t

args program command line arguments.

val envp : string array t

envp program environment variables.