Module Primus.Iterator

Iterator is a sequence of values of some domain.

Iterator is a just another abstraction that represent a sequence of values.

module type Base = sig ... end

Base interface of all iterators.

module type Finite = sig ... end

Finite iterators produce finite sequences of values.

module type Infinite = sig ... end

Infinite iterators produces infinite sequences.