Module type Iterator.Finite

Finite iterators produce finite sequences of values.

include Base

type t

iterator type

type dom

iterator domain

val min : dom

minimum value in the iterator domain

val max : dom

maximum value in the iterator domain

val value : t -> dom

current value

val next : t -> t option

next iterator moves an iterator to the next element of the sequence, or returns None if there are no more elements.