Module Attribute.Set

The set of attributes.

We represent the set of attributes as an abstracted value.

include Bap_core_theory.KB.Value.S with type t := set

include Ppx_sexp_conv_lib.Sexpable.S with type t := set

val t_of_sexp : Sexplib0__.Sexp.t -> set

val sexp_of_t : set -> Sexplib0__.Sexp.t

val empty : set

val domain : set Bap_knowledge.Knowledge.domain

include Base.Comparable.S with type t := set

val (>=) : set -> set -> bool

val (<=) : set -> set -> bool

val (=) : set -> set -> bool

val (>) : set -> set -> bool

val (<) : set -> set -> bool

val (<>) : set -> set -> bool

val equal : set -> set -> bool

val compare : set -> set -> int

val min : set -> set -> set

val max : set -> set -> set

val ascending : set -> set -> int

val descending : set -> set -> int

val between : set -> low:set -> high:set -> bool

val clamp_exn : set -> min:set -> max:set -> set

val clamp : set -> min:set -> max:set -> set Base__.Or_error.t

type comparator_witness

val comparator : ( set, comparator_witness ) Base__.Comparator.comparator

val validate_lbound : min:set Base__.Maybe_bound.t -> set Base__.Validate.check

val validate_ubound : max:set Base__.Maybe_bound.t -> set Base__.Validate.check

val validate_bound : min:set Base__.Maybe_bound.t -> max:set Base__.Maybe_bound.t -> set Base__.Validate.check

include Core_kernel.Binable.S with type t := set

val bin_size_t : set Bin_prot.Size.sizer

val bin_write_t : set Bin_prot.Write.writer

val bin_read_t : set Bin_prot.Read.reader

val __bin_read_t__ : ( int -> set ) Bin_prot.Read.reader

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : set Bin_prot.Type_class.writer

val bin_reader_t : set Bin_prot.Type_class.reader

val bin_t : set Bin_prot.Type_class.t

val get : 'a t -> set -> 'a

get attr set extracts the value of the attribute from the set.

val slot : ( Bap_core_theory.Theory.program, set ) Bap_core_theory.KB.slot

provides access to the definition's set of attributes.

This slot is provided when Semantics.enable function is called (it is called if the primus-lisp plugin is enabled.