Module Lisp.Closure

Machine independent closure.

A closure is an anonymous function, that performs some computation in the Machine Monad. Closures are used to extend the Lisp Machine with arbitrary primitive operations implemented in OCaml.

module type S = functor (Machine : Machine.S) -> sig ... end

type t = (module S)

module Make (Machine : Machine.S) : sig ... end