Module Lisp.Context

The set of constraints.

See the context section for more information.

Note, the context is automatically generated from the project, when a Primus Lisp program is loaded. But it is also possible to refine it using the unit's Lisp.Semantics.context slot.

type t = context

val create : (string * string list) list -> context

create [c1, f1;...; cN, fN] creates the context.

The created context will have classes c1,...,cN defined with feature sets f1,...,fN correspondingly.

val of_program : program -> context

of_project program the global program context.

val pp : Stdlib.Format.formatter -> context -> unit

Pretty-prints the context