Module Lisp.Message

Lisp Machine Message interface.

Lisp machine messages are sent with the msg primitive. The messages are abstract, but they could be printed.

type t = message

val pp : Stdlib.Format.formatter -> t -> unit

pp ppf msg prints the message into the specified formatter ppf.