Module Semantics.Value

Pure semantic value.

type t = value

val static : Bitvec.t -> t

static x a value statically equal to x.

val symbol : string -> t

symbol s a symbolic value.

A symbolic value is also static and the Primus Lisp lifter will intern symbols by asigning a static value for each symbol so that distinct symbols will have distinct static values.

val custom : (Bap_core_theory.Theory.Value.cls, 'a) Bap_core_theory.KB.slot -> 'a -> t

custom prop x creates a custom static value.

The property prop is set to x and the symbol slot is set to t, i.e., to true so that the value evaluates to non-nil in the condition expressions.

val nil : t

the false value that evaluates to nil symbol having 0 representation.