Module Lisp.Unit

A Primus Lisp module.

Primus Lisp programs are packed in modules in the knowledge base. The module has the Theory.Unit.source field that contains the program and other relevant fields. The source language is set to Unit.language.

since

2.3.0

val create : ?⁠name:string -> Bap_core_theory.Theory.Target.t -> Bap_core_theory.Theory.Unit.t Bap_core_theory.KB.t

create ?name target creates a Lisp unit.

The unit target field is set to target and the name is interned in the lisp package.

val is_lisp : Bap_core_theory.Theory.Unit.t -> bool Bap_core_theory.KB.t

is_lisp unit is true if unit was created with create.

val language : Bap_core_theory.Theory.language

The language identifier for Primus Lisp.