Module Make.Other

Local state of other machines.

This module gives access to the state of other machines. It is possible both to pry into others state and even to change their state to enable full intermachine communication. Use this module with care!

since

2.2.0

val get : id -> 'a state -> 'a t

get machine state returns the local state of the machine.

val put : id -> 'a state -> 'a -> unit t

put machine state sets the local state of the machine.

val update : id -> 'a state -> f:('a -> 'a) -> unit t

update machine state ~f maps the local state of the machine.