Module Value.Make

Make(Machine) provides an interface to the Value type lifted into the Machine monad.

Parameters

Machine : Machine.S

Signature

type t = value

type 'a m = 'a Machine.t

val id : t -> id

id x is a unique identifier of a value. Every evaluation of non-trivial computation produces a value with new identifier. Only setting and reading a variable preserves value identifiers. Each new constaint or arithmetic, or memory expression produces a value with a new identifier.

val to_word : t -> Bap.Std.word

to_word x projects x to a machine word. Note, many operations from the Word module are lifted into the Machine monad by this functor, so this operation is not usually necessary.

val of_word : Bap.Std.word -> t m

of_word x computes a fresh new value from x

val of_string : string -> t m

of_string s computes a fresh new value from a textual representation of a machine word x. See Bap.Std.Word module for more details.

val of_bool : bool -> t m

of_bool x creates a fresh new value from the boolean x.

val of_int : width:int -> int -> t m

of_int ~width x creates a fresh new value of the given width from the integer x

val of_int32 : ?⁠width:int -> int32 -> t m

of_int32 x creates a fresh new value from x

val of_int64 : ?⁠width:int -> int64 -> t m

of_int64 x creates a fresh new value from x

val b0 : t m

a fresh new false computation

val b1 : t m

a fresh new true computation

val one : int -> t m

one x same as of_word @@ one x

val zero : int -> t m

zero x same as of_word @@ zero x

val signed : t -> t m

signed x same as of_word @@ signed x

val is_zero : t -> bool

is_zero is lift1 Word.is_zero

val is_one : t -> bool

is_one is lift1 Word.is_one

val is_positive : t -> bool

is_positive is lift1 Word.is_positive

val is_negative : t -> bool

is_negative is lift1 Word.is_negative

val is_non_positive : t -> bool

is_non_positive is lift1 Word.is_non_positive

val is_non_negative : t -> bool

is_non_negative is lift1 Word.is_non_negative

val bitwidth : t -> int

bitwidth is lift1 Word.bitwidth

val extract : ?⁠hi:int -> ?⁠lo:int -> t -> t m

extracts ?hi ?lo is lift1 (Word.extract ?hi ?lo)

val concat : t -> t -> t m

concat is lift2 Word.concat

val succ : t -> t m

succ is lift1 Word.succ

val pred : t -> t m

pred is lift1 Word.pred

val nsucc : t -> int -> t m

nsucc see Word.nsucc

val npred : t -> int -> t m

npred see Word.npred

val abs : t -> t m

see Word.abs

val neg : t -> t m

see Word.neg

val add : t -> t -> t m

see Word.add

val sub : t -> t -> t m

see Word.sub

val mul : t -> t -> t m

see Word.mul

val div : t -> t -> t m

see Word.div

val modulo : t -> t -> t m

see Word.modulo

val (lnot) : t -> t m

see Word.lnot

val logand : t -> t -> t m

see Word.logand

val logor : t -> t -> t m

see Word.logor

val logxor : t -> t -> t m

see Word.logxor

val lshift : t -> t -> t m

see Word.lshift

val rshift : t -> t -> t m

see Word.rshift

val arshift : t -> t -> t m

see Word.arshift

module Syntax : sig ... end

Int-like syntax.

module Symbol : sig ... end

Symbol Value Isomorphism.

include Regular.Std.Regular.S with type t := t

type t

include Bin_prot.Binable.S with type t := t

type t

include Bin_prot.Binable.S_only_functions with type t := t

type t

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : (int -> t) Bin_prot.Read.reader

This function only needs implementation if t exposed to be a polymorphic variant. Despite what the type reads, this does *not* produce a function after reading; instead it takes the constructor tag (int) before reading and reads the rest of the variant t afterwards.

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_t : t Bin_prot.Type_class.t

include Ppx_sexp_conv_lib.Sexpable.S with type t := t

type t

val t_of_sexp : Sexplib0.Sexp.t -> t

val sexp_of_t : t -> Sexplib0.Sexp.t

val compare : t -> t -> int

include Regular.Std.Printable.S with type t := t

type t

type of printable

val to_string : t -> string

to_string x returns a human-readable representation of x

val str : unit -> t -> string

str () t is formatted output function that matches "%a" conversion format specifier in functions, that prints to string, e.g., sprintf, failwithf, errorf and, surprisingly all Lwt printing function, including Lwt_io.printf and logging (or any other function with type ('a,unit,string,...) formatN`. Example:

Or_error.errorf "type %a is not valid for %a"
  Type.str ty Exp.str exp

val pps : unit -> t -> string

synonym for str

val ppo : Core_kernel.Out_channel.t -> t -> unit

will print to a standard output_channel, useful for using in printf, fprintf, etc.

val pp_seq : Stdlib.Format.formatter -> t Core_kernel.Sequence.t -> unit

prints a sequence of values of type t

this will include pp function from Core that has type t printer, and can be used in Format.printf family of functions

include Core_kernel.Pretty_printer.S with type t := t

type t

val pp : Base.Formatter.t -> t -> unit

include Core_kernel.Comparable.S_binable with type t := t

include Core_kernel__.Comparable_intf.S_common

include Base.Comparable.S

include Base__.Comparable_intf.Polymorphic_compare

include Base.Comparisons.Infix

type t

val (>=) : t -> t -> bool

val (<=) : t -> t -> bool

val (=) : t -> t -> bool

val (>) : t -> t -> bool

val (<) : t -> t -> bool

val (<>) : t -> t -> bool

val equal : t -> t -> bool

val compare : t -> t -> int

compare t1 t2 returns 0 if t1 is equal to t2, a negative integer if t1 is less than t2, and a positive integer if t1 is greater than t2.

val min : t -> t -> t

val max : t -> t -> t

val ascending : t -> t -> int

ascending is identical to compare. descending x y = ascending y x. These are intended to be mnemonic when used like List.sort ~compare:ascending and List.sort ~cmp:descending, since they cause the list to be sorted in ascending or descending order, respectively.

val descending : t -> t -> int

val between : t -> low:t -> high:t -> bool

between t ~low ~high means low <= t <= high

val clamp_exn : t -> min:t -> max:t -> t

clamp_exn t ~min ~max returns t', the closest value to t such that between t' ~low:min ~high:max is true.

Raises if not (min <= max).

val clamp : t -> min:t -> max:t -> t Base.Or_error.t

include Base.Comparator.S with type t := t

type t

type comparator_witness

val comparator : (t, comparator_witness) Base.Comparator.comparator

include Base__.Comparable_intf.Validate with type t := t

type t

val validate_lbound : min:t Base.Maybe_bound.t -> t Base.Validate.check

val validate_ubound : max:t Base.Maybe_bound.t -> t Base.Validate.check

val validate_bound : min:t Base.Maybe_bound.t -> max:t Base.Maybe_bound.t -> t Base.Validate.check

module Replace_polymorphic_compare : Core_kernel__.Comparable_intf.Polymorphic_compare with type t := t

include Core_kernel__.Comparable_intf.Map_and_set_binable with type t := t with type comparator_witness := comparator_witness

type t

include Core_kernel.Comparator.S with type t := t

type t

type comparator_witness

val comparator : (t, comparator_witness) Core_kernel.Comparator.comparator

module Map : Core_kernel.Map.S_binable with type Key.t = t with type Key.comparator_witness = comparator_witness

module Set : Core_kernel.Set.S_binable with type Elt.t = t with type Elt.comparator_witness = comparator_witness

include Core_kernel.Hashable.S_binable with type t := t

type t

val hash_fold_t : Base.Hash.state -> t -> Base.Hash.state

val hash : t -> Base.Hash.hash_value

val hashable : t Core_kernel.Hashtbl.Hashable.t

module Table : Core_kernel.Hashtbl.S_binable with type key = t

module Hash_set : Core_kernel.Hash_set.S_binable with type elt = t

module Hash_queue : Core_kernel.Hash_queue.S with type Key.t = t

include Regular.Std.Data.S with type t := t

type t

type constructor

type info = string * [ `Ver of string ] * string option

name,Ver v,desc information attached to a particular reader or writer.

val version : string

Data representation version. After any change in data representation the version should be increased.

Serializers that are derived from a data representation must have the same version as a version of the data structure, from which it is derived. This kind of serializers can only read and write data of the same version.

Other serializers can actually read and write data independent on its representation version. A serializer, that can't store data of current version simply shouldn't be added to a set of serializers.

It is assumed, that if a reader and a writer has the same name and version, then whatever was written by the writer should be readable by the reader. The round-trip equality is not required, thus it is acceptable if some information is lost.

It is also possible, that a reader and a writer that has the same name are compatible. In that case it is recommended to use semantic versioning.

val size_in_bytes : ?⁠ver:string -> ?⁠fmt:string -> t -> int

size_in_bytes ?ver ?fmt datum returns the amount of bytes that is needed to represent datum in the given format and version

val of_bytes : ?⁠ver:string -> ?⁠fmt:string -> Regular.Std.bytes -> t

of_bytes ?ver ?fmt bytes deserializes a value from bytes.

val to_bytes : ?⁠ver:string -> ?⁠fmt:string -> t -> Regular.Std.bytes

to_bytes ?ver ?fmt datum serializes a datum to a sequence of bytes.

val blit_to_bytes : ?⁠ver:string -> ?⁠fmt:string -> Regular.Std.bytes -> t -> int -> unit

blit_to_bytes ?ver ?fmt buffer datum offset copies a serialized representation of datum into a buffer, starting from the offset.

val of_bigstring : ?⁠ver:string -> ?⁠fmt:string -> Core_kernel.bigstring -> t

of_bigstring ?ver ?fmt buf deserializes a datum from bigstring

val to_bigstring : ?⁠ver:string -> ?⁠fmt:string -> t -> Core_kernel.bigstring

of_bigstring ?ver ?fmt datum serializes a datum to a sequence of bytes represented as bigstring

val blit_to_bigstring : ?⁠ver:string -> ?⁠fmt:string -> Core_kernel.bigstring -> t -> int -> unit

blit_to_bigstring ?ver ?fmt buffer datum offset copies a serialized representation of datum into a buffer, starting from offset.

module Io : sig ... end

Input/Output functions for the given datum.

module Cache : sig ... end

Data cache.

val add_reader : ?⁠desc:string -> ver:string -> string -> t Regular.Std.reader -> unit

add_reader ?desc ~ver name reader registers a new reader with a provided name, version ver and optional description desc

val add_writer : ?⁠desc:string -> ver:string -> string -> t Regular.Std.writer -> unit

add_writer ?desc ~ver name writer registers a new writer with a provided name, version ver and optional description desc

val available_readers : unit -> info list

available_reader () lists available readers for the data type

val default_reader : unit -> info

default_reader returns information about default reader

val set_default_reader : ?⁠ver:string -> string -> unit

set_default_reader ?ver name sets new default reader. If version is not specified then the latest available version is used. Raises an exception if a reader with a given name doesn't exist.

val with_reader : ?⁠ver:string -> string -> (unit -> 'a) -> 'a

with_reader ?ver name operation temporary sets a default reader to a reader with a specified name and version. The default reader is restored after operation is finished.

val available_writers : unit -> info list

available_writer () lists available writers for the data type

val default_writer : unit -> info

default_writer returns information about the default writer

val set_default_writer : ?⁠ver:string -> string -> unit

set_default_writer ?ver name sets new default writer. If version is not specified then the latest available version is used. Raises an exception if a writer with a given name doesn't exist.

val with_writer : ?⁠ver:string -> string -> (unit -> 'a) -> 'a

with_writer ?ver name operation temporary sets a default writer to a writer with a specified name and version. The default writer is restored after operation is finished.

val default_printer : unit -> info option

default_writer optionally returns an information about default printer

val set_default_printer : ?⁠ver:string -> string -> unit

set_default_printer ?ver name sets new default printer. If version is not specified then the latest available version is used. Raises an exception if a printer with a given name doesn't exist.

val with_printer : ?⁠ver:string -> string -> (unit -> 'a) -> 'a

with_printer ?ver name operation temporary sets a default printer to a printer with a specified name and version. The default printer is restored after operation is finished.

Low level access to serializers

val find_reader : ?⁠ver:string -> string -> t Regular.Std.reader option

find_reader ?ver name lookups a reader with a given name. If version is not specified, then a reader with maximum version is returned.

val find_writer : ?⁠ver:string -> string -> t Regular.Std.writer option

find_writer ?ver name lookups a writer with a given name. If version is not specified, then a writer with maximum version is returned.