Module Primus.Info

Information about entities registered in the framework.

The Primus Framework is extended by registering in it certain entities, e.g., systems, components, observations. The list of registered items could be obtained and introspected using the info data structure.

The primus-systems, primus-components, and primus-observations commands provide the command line interface to this information.

since

2.1.0

val name : info -> Bap_knowledge.Knowledge.Name.t

the item name

val desc : info -> string

the item description

val long : info -> string

extended information about the item

val pp : Stdlib.Format.formatter -> info -> unit

prints the item