Module Primus.Job

A task to run a Primus system.

A is a system together with input parameters that is run via the Jobs module.

type t

an abstract type for jobs

val name : t -> string

the job name, doesn't have to be unique a bears any sense

val desc : t -> string

desc describes what the job is doing

val envp : t -> string array

an array of environment variables

val args : t -> string array

an array of execve parameters

val system : t -> system

the system that this job runs