Module type Linker.Code

Code representation.

A code representation is abstract and hides how the code itself is represented. It is just a function, that takes a machine and performs a computation using this machine.

Parameters

Machine : Machine.S

Signature

val exec : unit Machine.t

exec computes the code.