Module type Doc.Element

Abstract Element of a document.

A documentation element is something that can be printed. We keep it abstract, as we plan to extend it in the future.

type t

val pp : Stdlib.Format.formatter -> t -> unit