Module Lisp.Load

Primus Lisp program loader

type error

abstract error type

val program : ?⁠paths:string list -> Bap.Std.project -> string list -> (program, error) Core_kernel.result

program ?paths proj features loads a program that implements a set of features. For each feature its implementation file, that must have the same basename as the name of feature, is looked up in the list of directories, specified by the path parameter (defaults to the current folder). The first implementation that is found, will be used, thus the order of the paths matters.

Returns an abstract representation of a program, that can be linked into the Lisp machine, or an error if the program is not well-formed.

val pp_error : Stdlib.Format.formatter -> error -> unit

pp_error ppf err outputs error information into the pretty-printing formatter ppf.

val pp_program : Stdlib.Format.formatter -> program -> unit

pp_program ppf program dumps program definitions into the formatter ppf