Module type Machine.Component

A generic machine component.

A generic machine component is a functor, that is applied every time the Machine is instantiated. The init function is called when the Machine computation is started (the order in which components are initialized is not specified, but since all components store their state in the machine it doesn't matter).

The init function can perform any computation in the machine monad. But usually, it registers event observations.

Parameters

Machine : S

Signature

val init : unit -> unit Machine.t

init () component initialization function.