Module Memory.Descriptor

Abstract memory descriptor.

A desciptor uniquely identifies a memory device by its name. In addition, it holds meta information about memory address and data bus sizes.

type t = memory

val sexp_of_t : t -> Ppx_sexp_conv_lib.Sexp.t

val create : addr_size:int -> data_size:int -> string -> memory

create ~addr_size:m ~data_size:n name constructs a memory descriptor for a storage name with m lines in the address bus, and n bits in data.

val unknown : addr_size:int -> data_size:int -> memory

unknown ~addr_size:m ~data_size:n constructs a memory descriptor for an arbitrary storage with m lines in the address bus, and n bits in data.

val name : memory -> string

name memory returns memory identifier.

val addr_size : memory -> int

addr_size memory the number of bits in the address bus.

val data_size : memory -> int

data_size memory is the the number of bits in the data bus.

include Core_kernel.Comparable.S with type t := t

val (>=) : t -> t -> bool

val (<=) : t -> t -> bool

val (=) : t -> t -> bool

val (>) : t -> t -> bool

val (<) : t -> t -> bool

val (<>) : t -> t -> bool

val equal : t -> t -> bool

val compare : t -> t -> int

val min : t -> t -> t

val max : t -> t -> t

val ascending : t -> t -> int

val descending : t -> t -> int

val between : t -> low:t -> high:t -> bool

val clamp_exn : t -> min:t -> max:t -> t

val clamp : t -> min:t -> max:t -> t Base__.Or_error.t

type comparator_witness

val comparator : (t, comparator_witness) Base__.Comparator.comparator

val validate_lbound : min:t Base__.Maybe_bound.t -> t Base__.Validate.check

val validate_ubound : max:t Base__.Maybe_bound.t -> t Base__.Validate.check

val validate_bound : min:t Base__.Maybe_bound.t -> max:t Base__.Maybe_bound.t -> t Base__.Validate.check

module Replace_polymorphic_compare : sig ... end

module Map : sig ... end

module Set : sig ... end