Module Observation.Provider

Data interface to the provider.

This interface provides access to the data stream of all providers. The data stream is expressed as a stream of s-expressions.

type t = provider

val name : t -> string

the string representation of the provider name.

The name t is the textual representation of fullname t except that names in the package "primus" are printed unqualifed and all other names are qualified.

val fullname : t -> Bap_core_theory.KB.Name.t

the name of a provider.

val observers : t -> int

a total number of observers that subscribed to this provider

val triggers : t -> unit Bap_future.Std.stream

a stream of occurrences of this observation

val data : t -> Core_kernel.Sexp.t Bap_future.Std.stream

a data stream from this observation